requirements.txt
discord.py==2.3.2
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
requests==2.31.0
flask-sock==0.7.0
waitress==3.0.2
reportlab==4.4.10
pillow==12.1.1
